+++ Sources/Montage/3 Component/3 Selection And Input/Text/TextArea.swift
@@ -81,7 +81,7 @@ extension TextInput {
                 handler: (() -> Void)? = nil
             )
             case badge(
-                Badge.Content.Variant = .solid,
+                ContentBadge.Variant = .solid,
                 title: String
             )
             
@@ -445,7 +445,7 @@ extension TextInput {
                         handler: handler
                     )
                 case let .badge(variant, title):
-                    Badge.Content(variant: variant, text: title)
+                    ContentBadge(variant: variant, text: title)
                         .size(.medium)
                         .colorStyle(.neutral)
                 }
